Finished Chapter 13
let/cc is basically goto/exceptions/unwinding the stack as far as I can tell.

It really does seem out of place in a functional language but without
something like it there'd be no way to do many things in the most
efficient way possible and "error/solution" out early.